Drag & Drop Code Put Into Fragment B
In this commit, I added all the required  .xml files in order for the
drag and drop functionality to work. I have also changed/modified the
code to work in fragment b, no showing error's on the code, except when
ran it is throwing null pointer exceptions in the .onCreateView. Will
try and figure this issue out and then re commit later.